requirements.txt
python-dotenv>=1.0.0
python-telegram-bot>=20.3
requests>=2.31.0
pydantic>=2.6.0
faiss-cpu>=1.7.4
sentence-transformers>=2.5.1
google-api-python-client>=2.143.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
protobuf>=4.25.3
ujson>=5.9.0
pytest>=7.4.0
